outbound_group_session: Correctly free the session buffer.
This stores the buffer for the C lib inside the OlmOutboundGroupSession
struct.

While this technically violates aliasing rules since we give the
C lib a pointer to the vector, we never modify the vector avoiding any
reallocations.

We only need to make sure that the vector outlives the C group session.