account: Properly free the account buffer in the unpickling error case.
The buffer for the account was leaked using Box::into_raw(). The buffer
is only restored when we drop the OlmAccount object but if the
unpickling doesn't succeed we never construct an OlmAccount, thus we
leak the buffer.

This changes the way we store the buffer. The buffer is never boxed nor
leaked, the pointer is passed to the C lib and we make sure the buffer
stays alive as long as the OlmAccount object stays alive.