utility: Don't require the signature to be mutable.
The ed25519_verify() method of libolm doesn't mark the signature
argument as a const, this in turn tells bindgen that the argument should
be mutable.

The method in fact does not mutate the signature and there's no reason
for us to require a mutable reference to the signature.

We can work around this by casting the pointer to a mutable one.