olm-rs: Implement Send for all our structs.
Marking our structs as Send is safe because nothing will modify the pointer
under us from the C and it does not depend on any thread local storage,
the buffer that holds all the data is contained inside the struct.

Sync on the other hand is unsafe since libolm doesn't do any synchronization.