sas: Consume the public key when setting the other pubkey.
Since libolm clears the data in the buffer holding the base64 encoded
pubkey letting the lib user hold on to the string would mean that they
would end up with a unexpectedly garbled string.

So just consume it, this way the lib user will know that they need to
make a copy if they want to hold on to the public key.