Introduce a new ByteBuf type
and use it instead of Vec<u8> for buffers passed to libolm.
Now, safe code in this crate can no longer access the underlying buffer
of one of the libolm structs.

Additionally, the `Debug` implementation for `OlmSession` no longer
prints the internal buffer of the session object and buffers are no
longer zeroed before initialization by libolm.