Guard against crash on invalid base64 in `OlmInboundGroupSession`.
`olm_sys::olm_group_decrypt_max_plaintext_length` can return
`olm_error()` when passed an invalid base64 input. This value is a large
unsigned number, which is then passed as the dimension to `vec!`, making
the allocation fail and resulting in a crash.

To make this safe, we need to check the return value of
`olm_group_decrypt_max_plaintext_length` before passing it on.