Add a constructor to create an OlmPkDecryption object from a byte array
The OlmPkDecryption object is used to construct what the spec[1] calls a
Recovery key. This key can be backed up as a base 58 encoded string,
restoring the backup reverses the process. We get a byte array when we
decode the base 58 representation of the key.

This patch adds support to construct a OlmPkDecryption object from such
a byte array.

[1]: https://spec.matrix.org/v1.1/client-server-api/#recovery-key